# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.CreateResourceRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
      icon:
        type: string
      is_menu_visible:
        type: boolean
      key:
        type: string
      parent_id:
        type: string
      scope:
        type: string
      sort_order:
        type: integer
    required:
    - display_name
    - key
    - scope
    type: object
...
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
...
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.GrantRoleRequest:
    properties:
      academic_unit_id:
        type: string
      expires_at:
        type: string
      role_id:
        type: string
      school_id:
        type: string
    required:
    - role_id
    type: object
...
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ResourceDTO:
    properties:
      description:
        type: string
      display_name:
        type: string
      icon:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_menu_visible:
        type: boolean
      key:
        type: string
      parent_id:
        type: string
      scope:
        type: string
      sort_order:
        type: integer
    type: object
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.RoleDTO:
    properties:
      description:
        type: string
      display_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      scope:
        type: string
    type: object
...
  github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.UpdateResourceRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
      icon:
        type: string
      is_active:
        type: boolean
      is_menu_visible:
        type: boolean
      parent_id:
        type: string
      scope:
        type: string
      sort_order:
        type: integer
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Clean administration API for EduGo
  title: EduGo API Admin New
  version: "1.0"
paths:
...
  /resources:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ResourceDTO'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all resources
      tags:
      - resources
    post:
      consumes:
      - application/json
      parameters:
      - description: Resource data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.CreateResourceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ResourceDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a resource
      tags:
      - resources
  /resources/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ResourceDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a resource by ID
      tags:
      - resources
    put:
      consumes:
      - application/json
      parameters:
      - description: Resource ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Resource update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.UpdateResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ResourceDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a resource
      tags:
      - resources
  /roles:
    get:
      consumes:
      - application/json
      parameters:
      - description: Scope filter
        in: query
        name: scope
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.RoleDTO'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all roles
      tags:
      - roles
  /roles/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.RoleDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a role by ID
      tags:
      - roles
  /roles/{id}/permissions:
    get:
      consumes:
      - application/json
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get permissions for a role
      tags:
      - roles
...
  /users/{user_id}/roles:
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID (UUID)
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.RoleDTO'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get roles assigned to a user
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User ID (UUID)
        in: path
        name: user_id
        required: true
        type: string
      - description: Role grant data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.GrantRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.RoleDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Grant a role to a user
      tags:
      - users
  /users/{user_id}/roles/{role_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: User ID (UUID)
        in: path
        name: user_id
        required: true
        type: string
      - description: Role ID (UUID)
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduGoGroup_edugo-api-admin-new_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke a role from a user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT token. Example: "Bearer
      eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"